requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
betterpath==0.2.2
cffi==1.11.5
Click==7.0
cryptography==2.4.1
Django==2.1.2
django-cors-headers==2.4.0
django-filter==2.1.0
django-rest-knox==3.4.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
idna==2.7
l==0.11.0
Pillow==5.3.0
promise==2.2.1
pycparser==2.19
PyJWT==1.6.4
pyOpenSSL==18.0.0
pytz==2018.6
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
vcversioner==2.16.0.0
zope.interface==4.6.0
